RECORD 업로드 완료 시 GET API 설계

+++ src/main/java/com/nine/baseballdiary/backend/record/RecordRepository.java
@@ -2,5 +2,5 @@ package com.nine.baseballdiary.backend.record;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecordRepository extends JpaRepository<Record, Integer> {
+public interface RecordRepository extends JpaRepository<Record, Long> {
 }
